requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
gigachat==0.1.3
requests==2.32.5
pydantic==1.10.13